+++ src/fs/filesOperations/index.js
@@ -2,3 +2,4 @@ export * from './cat.js';
 export * from './add.js';
 export * from './rn.js';
 export * from './cp.js';
+export * from './rm.js';
